requirements.txt
openai
gradio
nltk
gtts
pathlib
python-dotenv